+++ app/src/renderer/vuex/modules/session.js
@@ -1,8 +1,6 @@
 import * as Sentry from "@sentry/browser"
 import noScroll from "no-scroll"
 import {
-  enableGoogleAnalytics,
-  disableGoogleAnalytics,
   track
 } from "../../google-analytics.js"
 import config from "../../../config"
@@ -39,8 +37,6 @@ export default () => {
       importKey,
       testPassword,
       generateSeed,
-      enableGoogleAnalytics,
-      disableGoogleAnalytics,
       track,
       Sentry
     }
@@ -208,19 +204,13 @@ export default () => {
           dsn: state.externals.config.sentry_dsn,
           release: state.externals.config.version
         })
-        state.externals.enableGoogleAnalytics(
-          state.externals.config.google_analytics_uid
-        )
-        console.log(`Analytics and error reporting have been enabled`)
+        console.log(`Error collection have been enabled`)
         state.externals.track(`pageview`, {
           dl: window.location.pathname
         })
       } else {
-        console.log(`Analytics disabled in browser`)
+        console.log(`Error collection disabled in browser`)
         state.externals.Sentry.init({})
-        state.externals.disableGoogleAnalytics(
-          state.externals.config.google_analytics_uid
-        )
       }
     }
   }